MBI-17: Setup Continuous Integration with Travis-CI (#3)
Continuous integration is the practice of merging all developer working copies to a shared 
mainline several times a day. When embarking on a change, a developer takes a copy of 
the current code base on which to work. As other developers submit changed code to 
the source code repository, this copy gradually ceases to reflect the repository code. 
Not only can the existing code base change, but new code can be added as well as new 
libraries, and other resources that create dependencies, and potential conflicts.

Implemented:

    — The continuous integration was integrated based on the Travis service that is the 
great solution for small libraries that should be tested and checked by linters, and 
deployed for cloud only. It required the configuration file .travis.tml. The integrated 
config contains the language, image of the MacOS and one command to tun the tests.

References:
    • Continuous integration — https://en.wikipedia.org/wiki/Continuous_integration
    • Travis CI — https://docs.travis-ci.com
    • Travis for iOS — https://docs.travis-ci.com/user/languages/objective-c